+++ tests/testthat/test-colwise-arrange.R
@@ -12,12 +12,17 @@ test_that(".funs is applied to variables before sorting", {
   expect_identical(arrange_all(df, `-`), arrange(df, -mpg, -cyl, -disp))
 })
 
-test_that("arrange_at can arrange by grouping variables (#3351)", {
+test_that("arrange_at can arrange by grouping variables (#3351, #3332)", {
   tbl <- data_frame(gr1 = rep(1:2, 4), gr2 = rep(1:2, each = 4), x = 1:8) %>%
     group_by(gr1)
 
   expect_identical(
     arrange_at(tbl, vars(gr1)),
     arrange(tbl, gr1)
+  )
+
+  expect_identical(
+    arrange_at(tbl, vars(-x)),
+    arrange(tbl, gr1, gr2)
   )
 })